fix: replace /home/page.tsx with its correct code from main.
